TightVNC 2.8.55
. Server and Viewer for Windows: Reverted recent changes in processing keyboard events that caused keyboard-related problems in version 2.8.53 (sf bugs #1530, #1531, #1532, #1533).
. Server and Viewer for Windows: Fixed a build problem that caused compatibility problems with certain versions of VNC software, due to using wrong protocol extension (sf bug #1534).
. Server and Viewer for Windows: Improved text labels and layout of elements in a number of screen forms.
. Server for Windows: Connecting to RDP sessions is now disabled by default and should be enabled explicitly, as sharing some virtual session instead of the main screen is probably not something a user would expect by default.
. Server for Windows: Fixed a minor resource leak (although it did not seem to cause any problems during normal operation).

+++ rfb/FrameBuffer.cpp
@@ -531,7 +531,7 @@ void *FrameBuffer::getBufferPtr(int x, int y) const
 
 int FrameBuffer::getBufferSize() const
 { 
-  return (m_dimension.area() * m_pixelFormat.bitsPerPixel) / 8;
+  return (int)((UINT64)m_dimension.area() * m_pixelFormat.bitsPerPixel / 8);
 }
 
 bool FrameBuffer::resizeBuffer()
